+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/src/routes/[lang]/about/+layout.server.ts
@@ -1,6 +1,6 @@
 import { initServerLoadWrapper } from "@inlang/sdk-js/adapter-sveltekit/server"
 import type { LayoutServerLoad } from "./$types.js"
 
-export const load = initServerLoadWrapper<LayoutServerLoad>().wrap(async (_, { i }) => {
+export const load = initServerLoadWrapper<LayoutServerLoad>().use(async (_, { i }) => {
 	console.info("[lang]/about/+layout.server.ts", i("welcome"))
 })