+++ source-code/sdk-js/src/adapter-sveltekit/runtime/client/shared/context.ts
@@ -1,42 +1,22 @@
 import { getContext } from "svelte"
 import { inlangSymbol } from "../../shared/utils.js"
+import { InlangSdkException } from "../../../vite-plugin/exceptions.js"
 
-// TODO: wrap with error message only during development
 export const getRuntimeFromContext = () => {
+	if (import.meta.env.PROD) {
+		return getContext(inlangSymbol)
+	}
+
+	// Showing the technical error message does only make sense in development
+	// and saves a few bytes from the production bundle
 	try {
 		return getContext(inlangSymbol)
-	} catch (e) {
-		throw Error(
-			`
-You cannot directly access any '@inlang/sdk-js' imports in this scope. You need to pass them from 'handle' or 'load' to the function you want to call e.g.
-
-// -- Change this -------------------------------------------------------------
-
-import { i } from '@inlang/sdk-js'
-
-export const load = async () => {
-   return { title: getPageTitle() }
-}
-
-const getPageTitle = () => {
-   console.log(i('hello.inlang'))
-}
-
-// -- To this -----------------------------------------------------------------
-
-import { i } from '@inlang/sdk-js'
-
-export const load = async () => {
-   return { title: getPageTitle(i) }
-}
-
-const getPageTitle = (/** @type { import('@inlang/sdk-js/runtime').InlangFunction } */ i) => {
-   console.log(i('hello.inlang'))
-}
-
-// ----------------------------------------------------------------------------
-`,
-			{ cause: e },
+	} catch (error) {
+		throw new InlangSdkException(
+			`You cannot directly access any '@inlang/sdk-js' imports in this scope.
+Please read the docs for more information on how to workaround this temporary limitation:
+https://inlang.com/documentation/sdk/sveltekit/advanced`,
+			error as Error,
 		)
 	}
 }