+++ source-code/plugins/json/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/plugin-json
 
+## 3.0.10
+
+### Patch Changes
+
+- 61ec4dac: chore: fix the id of the plugin from `inlang.plugin-i18next` to `inlang.plugin-json`
+
+## 3.0.9
+
+### Patch Changes
+
+- 0c82623d: fix ignore folder at getLanguage function
+
 ## 3.0.8
 
 ### Patch Changes